+++ src/main/java/com/woyao/domain/purchase/OrderItem.java
@@ -38,7 +38,7 @@ public class OrderItem extends DefaultModelImpl {
 	private Product product;
 
 	@Column(name = "UNIT_PRICE", nullable = false)
-	private long unitPrice;
+	private int unitPrice;
 
 	@Column(name = "QUANTITY", nullable = false)
 	private int quantity = 1;
@@ -72,11 +72,11 @@ public class OrderItem extends DefaultModelImpl {
 		this.product = product;
 	}
 
-	public long getUnitPrice() {
+	public int getUnitPrice() {
 		return unitPrice;
 	}
 
-	public void setUnitPrice(long unitPrice) {
+	public void setUnitPrice(int unitPrice) {
 		this.unitPrice = unitPrice;
 	}
